+++ package.json
@@ -16,6 +16,7 @@
     "test": "mocha && jshint ."
   },
   "dependencies": {
+    "bluebird": "^2.9.34",
     "co": "^4.5.4",
     "heroku-cli-util": "^5.3.0",
     "inflection": "^1.7.1",
